Move hasGetter/SetterSignature to utility class
